# software_SurtiTelas.Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SurtiTelas API",
    "version": "1.0.0",
    "description": "Backend API para la plataforma SurtiTelas - Manufactura y comercialización de prendas de vestir",
    "contact": {
      "name": "SurtiTelas Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "prod-123"
          },
          "ref": {
            "type": "string",
            "example": "REF-001"
          },
          "codigo": {
            "type": "string",
            "example": "CAM-001"
          },
          "nombre": {
            "type": "string",
            "example": "Camiseta básica"
          },
          "descripcion": {
            "type": "string",
            "example": "Camiseta 100% algodón"
          },
          "descripcionCorta": {
            "type": "string",
            "example": "Camiseta básica"
          },
          "categoria": {
            "type": "string",
            "example": "Camisetas"
          },
          "subcategoria": {
            "type": "string",
            "example": "Manga corta"
          },
          "marca": {
            "type": "string",
            "example": "SurtiTelas"
          },
          "precio": {
            "type": "number",
            "example": 25000
          },
          "precioAnterior": {
            "type": "number",
            "example": 30000
          },
          "descuento": {
            "type": "number",
            "example": 17
          },
          "cantidadStock": {
            "type": "number",
            "example": 100
          },
          "stock": {
            "type": "string",
            "enum": [
              "OK",
              "Bajo stock",
              "Agotado"
            ],
            "example": "OK"
          },
          "estado": {
            "type": "string",
            "enum": [
              "Activo",
              "Inactivo"
            ],
            "example": "Activo"
          },
          "imagenes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://cdn.surtitelas.com/img1.jpg"
            ]
          },
          "imagenPrincipal": {
            "type": "string",
            "example": "https://cdn.surtitelas.com/img1.jpg"
          },
          "publicado": {
            "type": "boolean",
            "example": true
          },
          "destacado": {
            "type": "boolean",
            "example": false
          },
          "oferta": {
            "type": "boolean",
            "example": true
          },
          "nuevo": {
            "type": "boolean",
            "example": false
          },
          "masVendido": {
            "type": "boolean",
            "example": true
          },
          "tela": {
            "type": "string",
            "example": "Algodón"
          },
          "colores": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Blanco",
              "Negro"
            ]
          },
          "tallas": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "S",
              "M",
              "L"
            ]
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "order-123"
          },
          "numero": {
            "type": "string",
            "example": "PED-000001"
          },
          "cliente": {
            "type": "string",
            "example": "Juan Pérez"
          },
          "asesor": {
            "type": "string",
            "example": "María Gómez"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-12T00:00:00.000Z"
          },
          "total": {
            "type": "number",
            "example": 50000
          },
          "items": {
            "type": "number",
            "example": 2
          },
          "estado": {
            "type": "string",
            "example": "Nuevo"
          },
          "prioridad": {
            "type": "string",
            "enum": [
              "Estándar",
              "Prioritario"
            ],
            "example": "Estándar"
          },
          "observaciones": {
            "type": "string",
            "example": "Entregar en la mañana"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-12T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-12T00:00:00.000Z"
          },
          "itemsList": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string",
                  "example": "prod-123"
                },
                "nombre": {
                  "type": "string",
                  "example": "Camiseta básica"
                },
                "precio": {
                  "type": "number",
                  "example": 25000
                },
                "cantidad": {
                  "type": "number",
                  "example": 2
                }
              }
            },
            "example": [
              {
                "productId": "prod-123",
                "nombre": "Camiseta básica",
                "precio": 25000,
                "cantidad": 2
              }
            ]
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "not_found"
          },
          "message": {
            "type": "string",
            "example": "Recurso no encontrado"
          }
        }
      },
...
      "Domiciliario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "dom-123"
          },
          "userId": {
            "type": "string",
            "example": "user-123"
          },
          "zona": {
            "type": "string",
            "nullable": true,
            "example": "Norte"
          },
          "vehiculo": {
            "type": "string",
            "nullable": true,
            "example": "Moto"
          },
          "capacidad": {
            "type": "integer",
            "nullable": true,
            "example": 50
          },
          "activo": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CmsPage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cms-123"
          },
          "slug": {
            "type": "string",
            "example": "inicio"
          },
          "titulo": {
            "type": "string",
            "example": "Inicio"
          },
          "contenido": {
            "type": "string",
            "nullable": true,
            "example": "<p>Bienvenido</p>"
          },
          "publicado": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateDomiciliario": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "example": "user-123"
          },
          "zona": {
            "type": "string",
            "nullable": true,
            "example": "Norte"
          },
          "vehiculo": {
            "type": "string",
            "nullable": true,
            "example": "Moto"
          },
          "capacidad": {
            "type": "integer",
            "nullable": true,
            "example": 50
          }
        }
      },
      "UpdateDomiciliario": {
        "type": "object",
        "properties": {
          "zona": {
            "type": "string",
            "nullable": true,
            "example": "Sur"
          },
          "vehiculo": {
            "type": "string",
            "nullable": true,
            "example": "Camioneta"
          },
          "capacidad": {
            "type": "integer",
            "nullable": true,
            "example": 100
          },
          "activo": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CreateCmsPage": {
        "type": "object",
        "required": [
          "slug",
          "titulo"
        ],
        "properties": {
          "slug": {
            "type": "string",
            "example": "inicio"
          },
          "titulo": {
            "type": "string",
            "example": "Inicio"
          },
          "contenido": {
            "type": "string",
            "nullable": true,
            "example": "<p>Bienvenido</p>"
          },
          "publicado": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "UpdateCmsPage": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "example": "inicio"
          },
          "titulo": {
            "type": "string",
            "example": "Inicio"
          },
          "contenido": {
            "type": "string",
            "nullable": true,
            "example": "<p>Bienvenido</p>"
          },
          "publicado": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/orders/{id}/domiciliario": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Assign delivery person to order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "domiciliarioId": "user-789"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delivery person assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Domiciliario asignado"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/domiciliarios": {
      "get": {
        "tags": [
          "Domiciliarios"
        ],
        "summary": "Lista domiciliarios",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "zona",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activo",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de domiciliarios"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "post": {
        "tags": [
          "Domiciliarios"
        ],
        "summary": "Crea un domiciliario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomiciliario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Domiciliario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Domiciliario"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/domiciliarios/{id}": {
      "get": {
        "tags": [
          "Domiciliarios"
        ],
        "summary": "Obtiene un domiciliario por ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domiciliario encontrado"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Domiciliarios"
        ],
        "summary": "Actualiza un domiciliario",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDomiciliario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Domiciliario actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Domiciliario"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
...
    "/cms": {
      "get": {
        "tags": [
          "CMS"
        ],
        "summary": "Lista páginas CMS",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de páginas CMS"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "post": {
        "tags": [
          "CMS"
        ],
        "summary": "Crea una página CMS",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCmsPage"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Página CMS creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CmsPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/cms/{id}": {
      "get": {
        "tags": [
          "CMS"
        ],
        "summary": "Obtiene una página CMS por ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página CMS encontrada"
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      },
      "patch": {
        "tags": [
          "CMS"
        ],
        "summary": "Actualiza una página CMS",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCmsPage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Página CMS actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CmsPage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/schemas/Error"
          },
          "403": {
            "$ref": "#/components/schemas/Error"
          },
          "404": {
            "$ref": "#/components/schemas/Error"
          }
        }
      }
    },
    "/catalog/products": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "List products with pagination",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string",
              "example": "Camiseta"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated products",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProducts"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Create product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "example": {
                "ref": "REF-001",
                "nombre": "Camiseta básica",
                "categoria": "Camisetas",
                "precio": 25000,
                "cantidadStock": 100,
                "stock": "OK",
                "estado": "Activo",
                "tela": "Algodón",
                "colores": [
                  "Blanco"
                ],
                "tallas": [
                  "M"
                ],
                "imagenes": [],
                "publicado": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Producto creado"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}